requirements.txt
snowflake-connector-python
pandas
pyarrow
pytest
python-dotenv
